@@ -1,6 +1,6 @@
 {
   "name": "@danielhokanson/ng-dynamic-forms",
-  "version": "19.0.3",
+  "version": "20.0.0",
   "description": "A rapid form development library for Angular",
   "repository": {
     "type": "git",
@@ -52,27 +52,28 @@
   },
   "private": false,
   "dependencies": {
-    "@angular/animations": "^19.0.0",
-    "@angular/cdk": "^19.0.0",
-    "@angular/common": "^19.0.0",
-    "@angular/compiler": "^19.0.0",
-    "@angular/core": "^19.0.0",
-    "@angular/forms": "^19.0.0",
-    "@angular/localize": "^19.0.0",
-    "@angular/material": "^19.0.0",
-    "@angular/platform-browser": "^19.0.0",
-    "@angular/platform-browser-dynamic": "^19.0.0",
-    "@angular/router": "^19.0.0",
-    "@ionic/angular": "^8.4.0",
-    "@ng-bootstrap/ng-bootstrap": "^18.0.0",
+    "@angular/animations": "^20.0.0",
+    "@angular/cdk": "^20.0.0",
+    "@angular/common": "^20.0.0",
+    "@angular/compiler": "^20.0.0",
+    "@angular/core": "^20.0.0",
+    "@angular/forms": "^20.0.0",
+    "@angular/localize": "^20.0.0",
+    "@angular/material": "^20.0.0",
+    "@angular/platform-browser": "^20.0.0",
+    "@angular/platform-browser-dynamic": "^20.0.0",
+    "@angular/router": "^20.0.0",
+    "@ionic/angular": "^8.7.0",
+    "@ng-bootstrap/ng-bootstrap": "^19.0.0",
+    "@primeuix/utils": "^0.6.3",
     "bootstrap": "^5.3.3",
     "core-js": "^3.39.0",
     "foundation-sites": "^6.9.0",
-    "ngx-bootstrap": "^19.0.2",
-    "ngx-mask": "^18.0.4",
+    "ngx-bootstrap": "^20.0.0",
+    "ngx-mask": "^20.0.0",
     "primeicons": "^7.0.0",
-    "primeng": "^19.0.0",
-    "quill": "^2.0.2",
+    "primeng": "^20.0.0",
+    "quill": "^2.0.3",
     "reflect-metadata": "^0.2.2",
     "rxjs": "^7.8.1",
     "tslib": "^2.8.1",
@@ -85,18 +86,18 @@
     "tmp": "^0.2.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^19.0.0",
-    "@angular-devkit/core": "^19.0.0",
-    "@angular-devkit/schematics": "^19.0.0",
-    "@angular-devkit/schematics-cli": "^19.0.0",
+    "@angular-devkit/build-angular": "^20.0.0",
+    "@angular-devkit/core": "^20.0.0",
+    "@angular-devkit/schematics": "^20.0.0",
+    "@angular-devkit/schematics-cli": "^20.0.0",
     "@angular-eslint/builder": "^20.6.0",
     "@angular-eslint/eslint-plugin": "^20.6.0",
     "@angular-eslint/eslint-plugin-template": "^20.6.0",
     "@angular-eslint/schematics": "^20.6.0",
     "@angular-eslint/template-parser": "^20.6.0",
-    "@angular/cli": "^19.0.0",
-    "@angular/compiler-cli": "^19.0.0",
-    "@angular/language-service": "^19.0.0",
+    "@angular/cli": "^20.0.0",
+    "@angular/compiler-cli": "^20.0.0",
+    "@angular/language-service": "^20.0.0",
     "@compodoc/compodoc": "^1.1.26",
     "@playwright/test": "^1.56.1",
     "@popperjs/core": "^2.11.8",
@@ -114,10 +115,10 @@
     "karma-coverage-istanbul-reporter": "^3.0.3",
     "karma-jasmine": "~5.1.0",
     "karma-jasmine-html-reporter": "^2.1.0",
-    "ng-packagr": "^19.0.0",
+    "ng-packagr": "^20.0.0",
     "size-limit": "^11.2.0",
     "ts-node": "~10.9.2",
-    "typescript": "~5.6.3"
+    "typescript": "~5.8.0"
   },
   "main": "index.js"
 }
